revert changes to tests in gh-24993
